requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
bs4==0.0.2
bs64==0.0.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
distro==1.9.0
emoji==2.15.0
fake-useragent==2.2.0
ffmpeg-python==0.2.0
frozenlist==1.8.0
future==1.0.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
multidict==6.7.0
munch==4.0.0
Naked==0.1.32
openai==2.6.0
pillow==12.0.0
propcache==0.4.1
psutil==7.1.1
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
pyfiglet==1.0.4
Pygments==2.19.2
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
setuptools==80.9.0
shellescape==3.8.1
sniffio==1.3.1
soupsieve==2.8
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
wheel==0.45.1
yarl==1.22.0
yt-dlp==2025.10.22
